Changes in QAndAForum Class Constructor Update:
Modified the constructor to accept an additional JFrame parameter (loginFrame), which is the login frame.
Stored this frame in a member variable to use later for logout functionality.
Added Logout Button:

Added a logoutButton to the UI and included it in the button panel.
Implemented an ActionListener for the logoutButton to dispose of the current frame and show the login frame.
Added Home Button:

Added a homeButton to the UI and included it in the button panel.
Implemented an ActionListener for the homeButton to reload questions by calling questionPanel.loadQuestions().